+++ src/types/continous-profiling.d.ts
@@ -1,4 +1,3 @@
-import { MonitorType } from "./../views/dashboard/related/continuous-profiling/data";
 /**
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
